Experimental improvement to SeoRedirect memory overhead
Instead of storing a dictionary of cached reflections to map between the
CallerFilePath/CallerLineNumber and the Controller/Action/Index state of
requests, now using a 64-bit hash of the the path/line number instead.

This will slightly increase computation at the cost of lower memory as the
variables in question are no longer stored as part of the lookup key. However,
it might increase lookup latency, but the change from using a string as the key
of the ConcurrentDictionary to a ulong should more than make up for this as
CityHash64 is far more efficient than the native hash.