Corrected exceptions triggered by optimizations breaking reflection
This was an expected issue, I was surprised to not have to deal with it sooner.
In the case of the initial SeoRedirect call wrapper and the main SeoRedirect
methods being combined/optimized/inlined/etc the stack frame is no longer the
second stack frame but the first.

Unfortunately in order to deal with the case when they are not inline as well,
and to keep the convenience methods available, it was necessary to break off
the code that obtains the calling method via reflection into each of the
SeoRedirect wrappers individually.